test: 예약 추가 테스트 결제 기능 DummyPaymentClient으로 대체
