+++ src/test/java/roomescape/service/ServiceTest.java
@@ -10,7 +10,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
+import roomescape.config.TestConfig;
 import roomescape.helper.CookieProvider;
 import roomescape.helper.DatabaseCleaner;
 import roomescape.helper.domain.MemberFixture;
@@ -20,6 +22,7 @@ import roomescape.helper.domain.ThemeFixture;
 import roomescape.helper.domain.WaitingFixture;
 
 @SpringBootTest(webEnvironment = WebEnvironment.NONE)
+@Import(TestConfig.class)
 @ActiveProfiles("test")
 abstract class ServiceTest {
     @Autowired
